postimg error correction

+++ fotorama_multi.php
@@ -20,21 +20,11 @@
 namespace mvbplugins\fotoramamulti;
 
 // fallback for wordpress security
-if (! defined('ABSPATH')) {
+if ( ! defined('ABSPATH' )) {
     die('Are you ok?');
 }
 
 const MAX_IMAGE_SIZE =  2560; // value for resize to ...-scaled.jpg TODO: big_image_size_threshold : read from WP settings. But where?
-
-require_once __DIR__ . '/inc/admin_settings.php';
-// run admin panel and check settings
-if (is_admin()) {
-	$fotorama_elevation = new FotoramaElevation();
-}
-
-// --------------- load additonal functions ------------------------------
-require_once __DIR__ . '/inc/fm_functions.php';
-require_once __DIR__ . '/languages/locales_i18n.php';
 
 // load globals and functions for status transitions only if needed or intended
 $const1 = get_option( 'fotorama_elevation_option_name' )['setCustomFields_15']; // liefert 'true' oder null
@@ -42,16 +32,15 @@ if ($const1 == 'true') {
 	require_once __DIR__ . '/inc/stateTransitions.php';
 }
 
-// load the wpseo_sitemap_url-images callback to add images of post to the sitemap only if needed or intended
-$yoastIsActivated = active( 'wordpress-seo/wp-seo.php' );
-$const2 = get_option( 'fotorama_elevation_option_name' )['doYoastXmlSitemap_16']; // liefert 'true' oder null
+require_once __DIR__ . '/inc/admin_settings.php';
 
-if ( ('true' == $const2) && $yoastIsActivated) {
-	require_once __DIR__ . '/inc/yoastXmlSitemap.php';
+// --------------- load additonal functions ------------------------------
+require_once __DIR__ . '/inc/fm_functions.php';
+
+if ( is_admin() ) {
+	$fotorama_elevation = new FotoramaElevation();
 }
 
-
-// -----------------------------------------------------------------------
 // define the shortcode to generate the image-slider with map
 add_shortcode('gpxview', '\mvbplugins\fotoramamulti\showmulti');
 
@@ -81,7 +70,7 @@ function showmulti($attr, $content = null)
 	// Extract shortcode-Parameters and set Default-Values
 	extract(shortcode_atts(array(
 		'gpxpath' => $fotorama_elevation_options['path_to_gpx_files_2'] ?? 'gpx',
-		'gpxfile' => '',
+		'gpxfile' => 'test.gpx',
 		'mapheight' => $fotorama_elevation_options['height_of_map_10'] ?? '450',
 		'chartheight' => $fotorama_elevation_options['height_of_chart_11'] ?? '200',
 		'imgpath' => $fotorama_elevation_options['path_to_images_for_fotorama_0'] ?? 'Bilder',
@@ -189,10 +178,10 @@ function showmulti($attr, $content = null)
 					'datesort' => $datesort, 'descr' => $description, 'thumbavail' => $thumbavail, 'thumbinsubdir' => $thumbinsubdir, 'alt' => $alt, 'caption' => $caption, 'sort' => $sort,
 				);
 			
-				// create array to add the image-urls to Yoast-seo xml-sitemap // TODO: anpassung für multi!!!
+				// create array to add the image-urls to Yoast-seo xml-sitemap
 				if ($doYoastXmlSitemap) {
 					$img2add = $up_url . '/' . $imgpath . '/' . $jpgfile . '.jpg';
-					$postimages[] = array('src' => $img2add , 'alt' => $title, );
+					$postimages[] = array('src' => $img2add , 'alt' => $title, 'title' => $title,);
 				}
 			
 				// increment imagenumber
@@ -211,13 +200,10 @@ function showmulti($attr, $content = null)
 		if ( ($rowsum != $arraysum) or ('true' == $ignoresort) ) {
 			$csort = array_column($data2, 'datesort');
 		}
-	}
-	
-	// sort images asending date-taken
-	if ($imageNumber > 0) {
+		// sort images asending with date-taken
 		array_multisort($csort, SORT_ASC, $data2);
 	}
-
+	
 	// on Status change from published to draft delete Custom-Fields 'lat' 'lon' and 'postimages' from the post
 	// delete always as we may have this plugin before and do want to delete the custom fields now. 
 	// But the deletion requires a status transition from published to draft and back
@@ -230,17 +216,30 @@ function showmulti($attr, $content = null)
 	// on the status transition of the post from 'draft' to 'published'.
 	// preset Custom-Field 'lat' and 'lon' of the post with GPS-Data of the first image 
 	// Will be overwritten with the first trackpoint of the GPX-track, if there is one provided
-	if ($draft_2_pub && (0 == $shortcodecounter) )  { // TODO: anpassung für multi!!!
-	//if (\current_user_can('edit_posts')) { 	
-		if ($setCustomFields) {
+	if ( \current_user_can('edit_posts') && $setCustomFields && (0 == $shortcodecounter)) {
 			gpxview_setpostgps($postid, $data2[0]['lat'], $data2[0]['lon']);
-		}
-		if ($doYoastXmlSitemap) { // TODO: append to existing postimg
-			$postimages = maybe_serialize($postimages);
-			delete_post_meta($postid,'postimg');
-			update_post_meta($postid,'postimg',$postimages,'');
-		}
 	}
+
+	if ( \current_user_can('edit_posts') && $setCustomFields && $doYoastXmlSitemap) { 
+			
+			if ( (0 == $shortcodecounter) ) {
+				delete_post_meta($postid,'postimg');
+
+			} else {
+				$myimgfrompost = get_post_meta($post_id,'postimg'); // read array with post_images from custom-field of post
+				
+				if ( ! empty($myimgfrompost) ) {
+					$test = $myimgfrompost[0]; // we need only the first index
+					$existimages = maybe_unserialize($test);	// type conversion to array
+					//append new images at the end of array
+					$postimages = \array_merge($existimages, $postimages);
+				}
+			}
+
+			$postimages = maybe_serialize($postimages);
+			update_post_meta( $postid, 'postimg', $postimages, '' );
+	}
+	
 			
 	// parse GPX-Track-Files, check if it is a file, and if so append to the string to pass to javascript
 	$files = explode(",", $gpxfile);
@@ -463,5 +462,109 @@ function showmulti($attr, $content = null)
 	return $htmlstring;
 }
 
-require_once __DIR__ . '/fotorama_multi_enq_scripts.php';
+//bind and call scripts and styles
+function fotomulti_scripts()
+{
+  wp_reset_query();
+  $plugin_url = plugins_url('/', __FILE__);
+  $useCDN = get_option( 'fotorama_elevation_option_name')['useCDN_13'];
+
+  if (!is_front_page() || !is_home()) {
+
+	// Load Styles
+	wp_enqueue_style('fm-style1', $plugin_url . 'css/fotorama_multi.css');
+	wp_enqueue_style('fm-style2', $plugin_url . 'css/fotorama3.min.css');
+	// Load Scripts
+	wp_enqueue_script('fm-script1', $plugin_url . 'js/fotorama3.min.js', array('jquery'), '3.1.0', true);
+	  
+	if ($useCDN == 'true') {
+		// Load Styles from CDN
+		wp_enqueue_style('fm-style3', 'https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.7.1/leaflet.min.css');
+		wp_enqueue_style('fm-style4', 'https://unpkg.com/@raruto/leaflet-elevation@1.5.3/dist/leaflet-elevation.min.css');
+		wp_enqueue_style('fm-style7', 'https://unpkg.com/leaflet-gesture-handling/dist/leaflet-gesture-handling.min.css');
+		wp_enqueue_style('fm-style5', 'https://cdnjs.cloudflare.com/ajax/libs/leaflet.fullscreen/2.0.0/Control.FullScreen.min.css');
+
+		// Load Scripts
+		wp_enqueue_script('fm-script3', 'https://cdnjs.cloudflare.com/ajax/libs/d3/5.16.0/d3.min.js', array('jquery'), '5.16.0', true); // does not work with d3 > version 6.0 !
+		wp_enqueue_script('fm-script2', 'https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.7.1/leaflet.min.js', array('jquery'), '1.7.1', true);
+		wp_enqueue_script('fm-script4', 'https://cdnjs.cloudflare.com/ajax/libs/leaflet-gpx/1.5.0/gpx.min.js', array('jquery'), '1.5.0', true);
+		wp_enqueue_script('fm-script12', $plugin_url . 'js/libs/leaflet-gpxgroup.min.js', array('jquery'), '', true);
+		wp_enqueue_script('fm-script5', 'https://unpkg.com/@raruto/leaflet-elevation@1.5.3/dist/leaflet-elevation.min.js', array('jquery'), '1.5.3', true);
+		wp_enqueue_script('fm-script8', 'https://unpkg.com/leaflet-gesture-handling/dist/leaflet-gesture-handling.min.js', array('jquery'), '', true);
+		wp_enqueue_script('fm-script10','https://cdnjs.cloudflare.com/ajax/libs/leaflet.fullscreen/2.0.0/Control.FullScreen.min.js', array('jquery'), '2.0.0', true);
+		wp_enqueue_script('fm-script11','https://cdnjs.cloudflare.com/ajax/libs/jquery-zoom/1.7.21/jquery.zoom.min.js', array('jquery'), '1.7.21', true);
+		wp_enqueue_script('fm-script9',  $plugin_url . 'js/fotorama_multi.min.js', array('jquery'), '', true);	
+	} else {
+		// Load local Styles
+		wp_enqueue_style('fm-style3', $plugin_url . 'css/leaflet.min.css');
+		wp_enqueue_style('fm-style4', $plugin_url . 'css/leaflet-elevation.min.css');
+		wp_enqueue_style('fm-style7', $plugin_url . 'css/leaflet-gesture-handling.min.css');
+		wp_enqueue_style('fm-style5', $plugin_url . 'css/Control.FullScreen.min.css');
+	
+		// Load local Scripts
+		wp_enqueue_script('fm-script3',  $plugin_url . 'js/libs/d3.min.js', array('jquery'), '5.15.0', true); // does not work with d3 > version 6.0 !
+		wp_enqueue_script('fm-script2',  $plugin_url . 'js/libs/leaflet.min.js', array('jquery'), '1.7.1', true);
+		wp_enqueue_script('fm-script4',  $plugin_url . 'js/libs/gpx.min.js', array('jquery'), '1.5.0', true);
+		wp_enqueue_script('fm-script12', $plugin_url . 'js/libs/leaflet-gpxgroup.min.js', array('jquery'), '', true);
+		wp_enqueue_script('fm-script5',  $plugin_url . 'js/libs/leaflet-elevation.min.js', array('jquery'), '1.5.3', true);
+		wp_enqueue_script('fm-script8',  $plugin_url . 'js/libs/leaflet-gesture-handling.min.js', array('jquery'), '', true);
+		wp_enqueue_script('fm-script10', $plugin_url . 'js/libs/Control.FullScreen.min.js', array('jquery'), '', true);
+		wp_enqueue_script('fm-script11', $plugin_url . 'js/zoom-master/jquery.zoom.min.js', array('jquery'), '1.7.21', true);
+		wp_enqueue_script('fm-script9',  $plugin_url . 'js/fotorama_multi.min.js', array('jquery'), '0.0.7', true);	
+		
+	}
+  }
+}
+
+add_action('wp_enqueue_scripts', 'mvbplugins\fotoramamulti\fotomulti_scripts');
+
+/**
+ * Prepare the translation of the back- and frontend
+ *
+ * @return void no return value
+ */
+function i18n_init() {
+	$dir = dirname( \plugin_basename( __FILE__)) . '/languages/';
+	load_plugin_textdomain( 'fotoramamulti', false, $dir);
+}
+
+add_action( 'plugins_loaded', 'mvbplugins\fotoramamulti\i18n_init'); // only for translations in the admin-settings page
+
+/**
+ * translate strings on client request (mind: it will not work if the page or post is cached by wordpress or another cache mechanism)
+ *
+ * @param $string $translate the string to translate
+ * @param $string $language	 the client language
+ * @return $string the translated string for defined language or the original string
+ */
+function t($translate, $language) {
+	
+	$languages = array('de', 'fr', 'it', 'es'); // provided translations
+
+	$de = array(
+		'Download' => 'Herunterladen',
+		'Start address' => 'Startadresse');
+
+	$fr = array(
+		'Download' => 'Télécharges',
+		'Start address' => 'Adresse de départ');
+
+	$it = array(
+		'Download' => 'Scarica',
+		'Start address' => 'Indirizzo iniziale');
+
+	$es = array(
+		'Download' => 'Descarga',
+		'Start address' => 'Dirección de inicio');
+
+	if ( ! in_array($language, $languages)) {
+		$language = "en";
+		return $translate;
+	} else {
+		$translate = isset( $$language[$translate]) ? $$language[$translate] : $translate;
+	}
+
+	return $translate;
+}
+
